venmo api, no login
venmo api is plugged in, not integrated yet, login is now no longer
mandatory in order to use the app. Checks will appear for account
specific actions (basically everything aside from scrolling the main
view) and will ask if you want to log in. The main function of this is
to let users sort of “demo” the app before committing to an account, so
less pressure to use. Also its a requirement for the App Store.